# Early exit for empty index in `vec_assign()`