@lionel- 2022-07-19
> Do you want an explicit test that vec_assign(x, integer(), -1L) works?

Yes please, it never hurts to add one more test :-)