@lionel- 2022-09-20
Closing for the reasons explained in https://github.com/r-lib/vctrs/pull/1636#issuecomment-1252300105