@hadley 2022-10-20
Reopening, because I'd like to reconsider in light of https://github.com/tidyverse/tidyr/pull/1392 — the performance gains seem meaningful, and it seems much more reasonable to handle this in vctrs rather than having to wrap every use of `vec_assign()`.